added admin manage child category

+++ app/Http/Controllers/Admin/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 
@@ -92,7 +93,14 @@ class SubCategoryController extends Controller
      */
     public function destroy(string $id)
     {
-         $subcat = SubCategory::findOrFail($id);
+        $subcat = SubCategory::findOrFail($id);
+        $childCat = ChildCategory::where(['sub_category_id' => $subcat->id])->count();
+
+        if ($childCat > 0) {
+            notyf()->error('There are child categories under this sub category. Please delete them first!');
+            return response(['status' => 'error']);
+        }
+
         $subcat->delete();
         notyf()->success('Sub Category Deleted Successfully!');
         return response(['status' => 'success']);
